+++ back/kickoff/src/main/java/server/kickoff/dto/MemberJoinDto.java
@@ -1,4 +1,4 @@
-package server.kickoff.repository;
+package server.kickoff.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
@@ -7,7 +7,7 @@ import lombok.Data;
 public class MemberJoinDto {
     @NotBlank(message = "아이디에 공백이 있으면 안됩니다.")
     private String userId;
-
+    @NotBlank(message = "비밀번호에 공백이 있으면 안됩니다.")
     private String password;
 
     private String userName;
